Update 03/12 - 03 - Bug fix that cause running incorrect generator
The original version only limits that at most one generator is active, but doesn't limit how many generator can wait. It may cause the program to run incorrect generator and may slow down the program.

To fix it, add a lock to set flag_stopGenerator, and let other thread return immediately if flag_stopGenerator is True, which means that there is already a thread waiting for stop.